md5.cpp: shut up 64-bit MSC warning.
svn:sync-xref-src-repo-rev: r39104

+++ src/VBox/Runtime/common/checksum/md5.cpp
@@ -1,4 +1,4 @@
-/* $Id: md5.cpp 8155 2008-04-18 15:16:47Z noreply@oracle.com $ */
+/* $Id: md5.cpp 14048 2008-11-10 22:43:16Z knut.osmundsen@oracle.com $ */
 /** @file
  * MD5 message digest functions
  */
@@ -111,7 +111,7 @@ RTDECL(void) RTMd5Update(PRTMD5CONTEXT ctx, const void *pvBuf, size_t len)
     t = ctx->bits[0];
     if ((ctx->bits[0] = t + ((uint32_t) len << 3)) < t)
 	ctx->bits[1]++; 	/* Carry from low to high */
-    ctx->bits[1] += len >> 29;
+    ctx->bits[1] += (uint32_t)(len >> 29);
 
     t = (t >> 3) & 0x3f;	/* Bytes already in shsInfo->data */
